core/agentic_ai/requirements.txt
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
streamlit>=1.28.0
langgraph>=0.0.15
wikipedia>=1.4.0
speechrecognition>=3.10.0
